# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Model Eval Console",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/candidates": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Register Candidate",
        "operationId": "register_candidate_v1_candidates_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "List Candidates",
        "operationId": "list_candidates_v1_candidates_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateOut"
                  },
                  "title": "Response List Candidates V1 Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/datasets": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "List Datasets",
        "operationId": "list_datasets_v1_datasets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetOut"
                  },
                  "type": "array",
                  "title": "Response List Datasets V1 Datasets Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Ingest Dataset",
        "operationId": "ingest_dataset_v1_datasets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Enqueue Run",
        "operationId": "enqueue_run_v1_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_v1_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunOut"
                  },
                  "title": "Response List Runs V1 Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/predictions": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Predictions",
        "operationId": "predictions_v1_runs__run_id__predictions_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "correct",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Correct"
            }
          },
          {
            "name": "slice",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Slice"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PredictionOut"
                  },
                  "title": "Response Predictions V1 Runs  Run Id  Predictions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/comparisons": {
      "get": {
        "tags": [
          "comparisons"
        ],
        "summary": "List Comparisons",
        "operationId": "list_comparisons_v1_comparisons_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComparisonOut"
                  },
                  "title": "Response List Comparisons V1 Comparisons Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "comparisons"
        ],
        "summary": "Compare Runs",
        "operationId": "compare_runs_v1_comparisons_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparisonCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/comparisons/{comparison_id}": {
      "get": {
        "tags": [
          "comparisons"
        ],
        "summary": "Read Comparison",
        "operationId": "read_comparison_v1_comparisons__comparison_id__get",
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Comparison Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CandidateCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[\\w .-]+$",
            "title": "Name"
          },
          "hf_model_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[\\w.-]+(/[\\w.-]+)?$",
            "title": "Hf Model Id"
          },
          "hf_revision": {
            "type": "string",
            "pattern": "^[0-9a-f]{40}$",
            "title": "Hf Revision"
          },
          "serving_config": {
            "$ref": "#/components/schemas/ServingConfig"
          }
        },
        "type": "object",
        "required": [
          "name",
          "hf_model_id",
          "hf_revision"
        ],
        "title": "CandidateCreate"
      },
      "CandidateOut": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[\\w .-]+$",
            "title": "Name"
          },
          "hf_model_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[\\w.-]+(/[\\w.-]+)?$",
            "title": "Hf Model Id"
          },
          "hf_revision": {
            "type": "string",
            "pattern": "^[0-9a-f]{40}$",
            "title": "Hf Revision"
          },
          "serving_config": {
            "$ref": "#/components/schemas/ServingConfig"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "hf_model_id",
          "hf_revision",
          "serving_config",
          "id",
          "created_at"
        ],
        "title": "CandidateOut"
      },
      "ComparisonCreate": {
        "properties": {
          "baseline_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Baseline Run Id"
          },
          "candidate_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Candidate Run Id"
          },
          "metric": {
            "type": "string",
            "const": "accuracy",
            "title": "Metric",
            "default": "accuracy"
          },
          "margin": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Margin",
            "default": 0.01
          },
          "n_resamples": {
            "type": "integer",
            "maximum": 50000.0,
            "minimum": 100.0,
            "title": "N Resamples",
            "default": 10000
          },
          "seed": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Seed",
            "default": 42
          }
        },
        "type": "object",
        "required": [
          "baseline_run_id",
          "candidate_run_id"
        ],
        "title": "ComparisonCreate"
      },
      "ComparisonOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "baseline": {
            "$ref": "#/components/schemas/ScoreRef"
          },
          "candidate": {
            "$ref": "#/components/schemas/ScoreRef"
          },
          "margin": {
            "type": "number",
            "title": "Margin"
          },
          "delta": {
            "type": "number",
            "title": "Delta"
          },
          "ci95": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Ci95"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "pass",
              "fail",
              "inconclusive"
            ],
            "title": "Verdict"
          },
          "mcnemar_p": {
            "type": "number",
            "title": "Mcnemar P"
          },
          "regressions": {
            "type": "integer",
            "title": "Regressions"
          },
          "fixes": {
            "type": "integer",
            "title": "Fixes"
          },
          "standard_error": {
            "type": "number",
            "title": "Standard Error"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "slices": {
            "items": {
              "$ref": "#/components/schemas/SliceResult"
            },
            "type": "array",
            "title": "Slices"
          },
          "flips": {
            "items": {
              "$ref": "#/components/schemas/Flip"
            },
            "type": "array",
            "title": "Flips"
          },
          "near_boundary_share": {
            "type": "number",
            "title": "Near Boundary Share"
          },
          "n_examples": {
            "type": "integer",
            "title": "N Examples"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "n_resamples": {
            "type": "integer",
            "title": "N Resamples"
          },
          "latency_budget_pass": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Budget Pass"
          },
          "latency_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ratio"
          },
          "evidence": {
            "type": "string",
            "enum": [
              "measured",
              "fixture"
            ],
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "metric",
          "baseline",
          "candidate",
          "margin",
          "delta",
          "ci95",
          "verdict",
          "mcnemar_p",
          "regressions",
          "fixes",
          "standard_error",
          "warnings",
          "slices",
          "flips",
          "near_boundary_share",
          "n_examples",
          "seed",
          "n_resamples",
          "latency_budget_pass",
          "latency_ratio",
          "evidence"
        ],
        "title": "ComparisonOut"
      },
      "DatasetCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "default": "SST-2 validation"
          },
          "hf_dataset": {
            "type": "string",
            "enum": [
              "stanfordnlp/sst2",
              "stanfordnlp/imdb"
            ],
            "title": "Hf Dataset",
            "default": "stanfordnlp/sst2"
          },
          "hf_config": {
            "type": "string",
            "enum": [
              "default",
              "plain_text"
            ],
            "title": "Hf Config",
            "default": "default"
          },
          "split": {
            "type": "string",
            "enum": [
              "validation",
              "test"
            ],
            "title": "Split",
            "default": "validation"
          },
          "hf_revision": {
            "type": "string",
            "pattern": "^[0-9a-f]{40}$",
            "title": "Hf Revision"
          }
        },
        "type": "object",
        "required": [
          "hf_revision"
        ],
        "title": "DatasetCreate"
      },
      "DatasetOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "hf_dataset": {
            "type": "string",
            "title": "Hf Dataset"
          },
          "hf_config": {
            "type": "string",
            "title": "Hf Config"
          },
          "split": {
            "type": "string",
            "title": "Split"
          },
          "hf_revision": {
            "type": "string",
            "title": "Hf Revision"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "n_examples": {
            "type": "integer",
            "title": "N Examples"
          },
          "provenance": {
            "additionalProperties": true,
            "type": "object",
            "title": "Provenance"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "hf_dataset",
          "hf_config",
          "split",
          "hf_revision",
          "content_hash",
          "n_examples"
        ],
        "title": "DatasetOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PredictionOut": {
        "properties": {
          "idx": {
            "type": "integer",
            "title": "Idx"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "label": {
            "type": "integer",
            "title": "Label"
          },
          "pred": {
            "type": "integer",
            "title": "Pred"
          },
          "prob_pos": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prob Pos"
          },
          "raw_output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Output"
          },
          "correct": {
            "type": "boolean",
            "title": "Correct"
          },
          "slices": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Slices"
          }
        },
        "type": "object",
        "required": [
          "idx",
          "text",
          "label",
          "pred",
          "prob_pos",
          "correct",
          "slices"
        ],
        "title": "PredictionOut"
      },
      "RunCreate": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "format": "uuid",
            "title": "Candidate Id"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "dataset_id"
        ],
        "title": "RunCreate"
      },
      "RunOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "candidate_id": {
            "type": "string",
            "format": "uuid",
            "title": "Candidate Id"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "max_attempts": {
            "type": "integer",
            "title": "Max Attempts"
          },
          "code_version": {
            "type": "string",
            "title": "Code Version"
          },
          "env": {
            "additionalProperties": true,
            "type": "object",
            "title": "Env"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "default": 0
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "candidate_name": {
            "type": "string",
            "title": "Candidate Name",
            "default": ""
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "candidate_id",
          "dataset_id",
          "status",
          "attempts",
          "max_attempts",
          "code_version",
          "env",
          "metrics",
          "error",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "RunOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}